pipeline/requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
# moomoo enrichment (local only — not used in GitHub Actions)
moomoo-openapi>=6.0.0
supabase>=2.0.0
